move files to src folder
